fix content and router
